default to current file selection
