+++ core/protocol/ideWebview.ts
@@ -51,6 +51,7 @@ export type ToIdeFromWebviewProtocol = ToIdeFromWebviewOrCoreProtocol & {
     },
     string | undefined,
   ];
+  "edit/addCurrentSelection": [undefined, void];
   "edit/clearDecorations": [{ shouldFocusEditor: boolean }, void];
 };
